+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Probability for dummies".
 - Tests for "Tricky number verifier".
 - Tests for "Crop-circles".
+- Tests for "Nicholas Breakspeare and Hugh of Evesham".
 
 ### Changed
 - Renaming "Linear Bézier curves" to "Cubic Bézier curves".